requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
ariadne-codegen==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
astroid==3.3.9 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4.0"
autoflake==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
black==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==44.0.2 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
dill==0.3.9 ; python_version >= "3.12" and python_version < "4.0"
firecloud==0.16.38 ; python_version >= "3.12" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.12" and python_version < "4"
functions-framework==3.8.2 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.24.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.6.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.4 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-secret-manager==2.23.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.19.0 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.69.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.69.2 ; python_version >= "3.12" and python_version < "4.0"
graphql-core==3.2.6 ; python_version >= "3.12" and python_version < "4"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.71.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.71.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4" and platform_system != "Windows"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
hypothesis==6.131.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.12" and python_version < "4.0"
isort==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4"
mccabe==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
nebelung==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
nose==1.3.7 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.2.4 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.32.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pandera[strategies]==0.22.1 ; python_version >= "3.12" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
pd-flatten==0.1.1 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.3.7 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.4 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.3 ; python_version >= "3.12" and python_version < "4.0"
pydot==3.0.4 ; python_version >= "3.12" and python_version < "4.0"
pyflakes==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
pygithub==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.12" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pylint==3.3.6 ; python_version >= "3.12" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
python-baseconv==1.2.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
requests[security]==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rich==14.0.0 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
setuptools==78.1.0 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
typeguard==4.4.2 ; python_version >= "3.12" and python_version < "4.0"
typer==0.15.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.1.3 ; python_version >= "3.12" and python_version < "4"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
